:recycle: Separate file for config reader
