+++ Sources/Kyusu/Reader/YamlReader.swift
@@ -8,18 +8,6 @@
 import Foundation
 import Yaml
 
-public typealias Path = String
-public typealias Command = [String]
-public struct YamlConfig {
-    public let sourcePaths: [Path]
-    public let ignoredPaths: [Path]
-    public let command: Command
-}
-
-public protocol ConfigReader {
-    associatedtype Config
-    func read(filePath: Path) throws -> Config
-}
 
 public struct YamlConfigReader: ConfigReader {
     public init() {
